+++ substance/SubstancePainter/Scripts/Prism_SubstancePainter_ExportTextureUI.py
@@ -1,11 +1,12 @@
-from qtpy.QtCore import *
-from qtpy.QtGui import *
-from qtpy.QtWidgets import *
+from PySide6.QtCore import *
+from PySide6.QtGui import *
+from PySide6.QtWidgets import *
 
-from qtpy.QtCore import Qt
+from PySide6.QtCore import Qt
 import sys
 
 import substance_painter.textureset
+import substance_painter.ui
 
 # Get the type used by flags in this Qt binding
 FLAG_TYPE = type(QTreeWidgetItem().flags())
@@ -30,7 +31,7 @@ ITEM_IS_SELECTABLE = _get_flag("ItemIsSelectable", 0x00000001)
 
 
 class TextureExportUI(QDialog):
-    def __init__(self, parent=None):
+    def __init__(self, parent):
         super().__init__(parent)
         self.setWindowTitle("Texture Export")
         self.resize(600, 500)
@@ -159,9 +160,5 @@ class TextureExportUI(QDialog):
 
         return textureData
 
-
-if __name__ == "__main__":
-    app = QApplication(sys.argv)
-    window = TextureExportUI()
-    window.show()
-    sys.exit(app.exec_())
+    def __del__(self):
+        print("Texture Export UI has been deleted !")
