--- substance/SubstancePainter_tester/Scripts/Prism_SubstancePainter_tester_externalAccess_Functions.py
@@ -1,68 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-####################################################
-#
-# PRISM - Pipeline for animation and VFX projects
-#
-# www.prism-pipeline.com
-#
-# contact: contact@prism-pipeline.com
-#
-####################################################
-#
-#
-# Copyright (C) 2016-2023 Richard Frangenberg
-# Copyright (C) 2023 Prism Software GmbH
-#
-# Licensed under GNU LGPL-3.0-or-later
-#
-# This file is part of Prism.
-#
-# Prism is free software: you can redistribute it and/or modify
-# it under the terms of the GNU Lesser General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Prism is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU Lesser General Public License for more details.
-#
-# You should have received a copy of the GNU Lesser General Public License
-# along with Prism.  If not, see <https://www.gnu.org/licenses/>.
-
-
-import os
-import platform
-
-from qtpy.QtCore import *
-from qtpy.QtGui import *
-from qtpy.QtWidgets import *
-
-from PrismUtils.Decorators import err_catcher_plugin as err_catcher
-
-
-class Prism_SubstancePainter_tester_externalAccess_Functions(object):
-    def __init__(self, core, plugin):
-        self.core = core
-        self.plugin = plugin
-
-    @err_catcher(name=__name__)
-    def getAutobackPath(self, origin):
-        autobackpath = ""
-        if platform.system() == "Windows":
-            autobackpath = os.path.join(
-                self.core.getWindowsDocumentsPath(), "SubstancePainter_tester"
-            )
-
-        fileStr = "SubstancePainter_tester Scene File ("
-        for i in self.sceneFormats:
-            fileStr += "*%s " % i
-
-        fileStr += ")"
-
-        return autobackpath, fileStr
-
-    @err_catcher(name=__name__)
-    def copySceneFile(self, origin, origFile, targetPath, mode="copy"):
-        pass
